Add second buffer for partial bounding boxes detection
Add a second layer buffer for partial bounding box detection, currently it only activate when bounding box is near the bottom of the screen (720>y>710).
It's not an overall imrpovement on every cameras and every date, might need to consider which camera or date to use this functionality.
The threshold of second buffer could be adjusted, maybe it will work better with our selected parameters.